Improve Code DB-NOTIFICATION-CONTROLLER

+++ database/migrations/2020_02_18_123139_create_post_table.php
@@ -34,8 +34,8 @@ class CreatePostTable extends Migration
             // ->on('user_profiles')
             // ->onDelete('cascade');
 
-             $table->increments('id');
-            $table->integer('user_profile_id')->unsigned();
+             $table->bigIncrements('id');
+            $table->bigInteger('user_profile_id')->unsigned();
               $table->enum('type', config('constants.enums.post_type'));
             $table->string('text')->nullable();
             $table->enum('alignment',['left','center','right'])->nullable();
