+++ database/migrations/2020_02_26_172925_create_followers_table.php
@@ -14,9 +14,9 @@ class CreateFollowersTable extends Migration
     public function up()
     {
          Schema::create('followers', function (Blueprint $table) {
-        $table->increments('id');
-        $table->integer('follower_id')->unsigned();
-        $table->integer('user_id')->unsigned();
+        $table->bigIncrements('id');
+        $table->bigInteger('follower_id')->unsigned();
+        $table->bigInteger('user_id')->unsigned();
         $table->timestamps();
 
         $table->foreign('follower_id')->references('id')->on('users')->onDelete('cascade');
